CV Code
I added the first iteration of the CV code.
This is capable of finding the closest enemy and returning a vector from the center of the camera to the center of the enemy based on the pixel coordinates of the camera.
Next version will be able to calculate world coordinates and detect if the enemy is moving.